added powerups
I've added powerups as one of the required families for the first build. There's one that heals, one that increases attack, one that increases speed, one that gives a shield, and one that gives a multishot.
I've also implemented a TileMap, so that you don't have to manually insert each tile individually. This is only good for walls though, as it also has a collider. Another TileMap will have to be added for "background" tiles (flooring, etc.). The player still phases through the wall.
I've also generalized the health script. It should be applicable to anything that has health (players, enemies, shields, etc.).