added title screen
I've added a title screen, as required for the first build.
The majourity of this update through is that I've reformatted the powerups, so that they all derive from a common class (like the object tree pattern). The multishot powerup is loosely connected to the tree; it'll have to be properly reformed later.
I've also generalized the bullet script, so that it can be reused for enemy attacks.