fixed camera
I've added a camera similar to the one from the Tanks tutorial; it follows the two players around and zooms in and out as needed.
I think the problem before was that the main camera was focusing on the most recent camera created: the second player's camera. I just removed the players' cameras and created a general camera.
Once we get enemy bullets up, we should test whether or not the camera snaps to one player if the other dies.