added player ui
I've added a canvas to the camera, so that the UI follows the camera around. The second player UI disappears if there's only one player.
Still need to test if the player's health + UI changes if hit with a damage source. Also, the ammo bar is currently nonfunctional, since there's nothing in-game that's keeping track of ammo amount.