added shield ui
I forgot to add the shield UI last time, so it's in here now.
I've changed the way the shield works, so that it would be easier for the UI to track. Instead of the shield powerup spawning a shield that follows the player, now the player has the shield as its child (though it starts off inactive). When the player gets a shield powerup, the shield activates, and when the shield runs out of health, the shield deactivates.
I've also reformatted the canvas manager, so that it works using a list. This is just for ease of use, that way you don't have to specify "player one" or "player two" within the variable names. You still need to know the index to access things though. I'd suggest looking at the code, because I don't know how to explain this.