added (some) controller support
The second player can now move using a controller. I'm not sure if the control scheme will work with other controllers, as I only tested it with mine.
There is no aiming implemented yet, the second player still focuses on the cursor.
I also left in the original arrow-key scheme.