added enemy AI and updated enemy shooting
Added enemy movement. The enemy has a "sight" collider, so any player that enters the collider, the enemy will begin shooting and pursuing the player set as target.
I edited the enemy shooting so now the enemy shoots at the player's location.
There's a problem, the player bullets hits the "sight" collider and depletes enemy health.