Updated enemy AI, added enemy health script and enemy UI
I updated the enemy detection. The enemy will pursue the player to a designated range, when the player gets too close the enemy will back up and keep shooting.

I added a enemy health script so I could display the enemy UI. It's pretty much the same as the health script with added code for the UI.

I updated the bulletHit script so now the bullet will differentiate from the regular health script and the enemy health script.