reimplemented the powerup ui (again)
I've reimplemented the powerup UI again, and also fixed the issue with the multishot powerup icon not displaying correctly. I've also reformatted the multishot powerup to be more connected with the powerup base.
I've created an extra scene for testing powerups and the UI.
I've also implemented the new player sprite. Right now, it only shows the front, so implementing the other sides should be next. (Reminder, change the player's colliders if you're changing their sprite.)