added working title screen
The game now starts on a title screen, where the player can either choose to enter the game, or to quit the application. I also fixed the issue with the buttons on the title screen not working. (Don't delete the EventSystem.)
I also moved the spawn points of the players on the SampleScene; they spawn within enemy view, so the enemy immediately attacks them.