added game over screen
I added a game over screen, which is overlayed on top of the gameplay screen, once all players are dead.
I also fixed the enemy scripts and the UI-related scripts, so that there aren't excessive exception messages in the console (still have no idea how to fix that starting camera exception though).