added (partial) game loop
The game over screen now has a button that leads the player back to the title screen. This partially satisfies the game loop aspect, but we still need a win condition.
I've also restructured the hierarchy to make it look neater.
Also, I fixed the UI, because it was incorrectly displaying the game over screen, and wasn't displaying the player's UIs.