added two new enemies
I added the explosion enemy, which shows a countdown before it explodes, and the burst enemy, which shoots a ring of bullets once it sees the player.
I've also generalized some scripts for efficiency, and reordered some things.
I'm a little unhappy with how the explosion is handled. I can't figure out how to smoothly push the player back from the explosion; currently it looks stiff and unnatural.
Aside from that, the only thing we have left to do is to create a level and a win condition.
(Also, I finally fixed the player's wall-collision absurdities. Turns out it was because the player didn't have a dynamic rigidbody. I fixed the enemies' collisions with the wall as well.)