Added the 2D NavMesh package
Added the NavMesh assets for 2D scenes from canvas. I added a ground and walls to the single player scene and added a NavMesh object. Also added the NavMesh agents to the enemies and the player.

I think the controls are a little buggy, also we need to work on the player detection code, for now the enemies can detect the player behind the walls.