fixed second player aiming, other small things
I fixed the second player's aiming (and reworked the aiming code overall) so that the second player no longer relies on the mouse to aim. I've also added reticles to the players, so they know where they're aiming.
I've also added a credits scene that displays everyone's names. Since I didn't want to order them in a particular way, the order randomizes every time the scene is opened.
I've also added a barrier in the player select scene; it prevents player from choosing two-player mode if the game doesn't detect any controllers plugged in. Not sure if this works with Parsec though, so this will need remote testing.