fixed movement-after-win bug; started fixing encapsulation
One of the playtesters reported a bug where the player would continue to move after winning the game. That's been fixed, so now the player just stops in their tracks once they win.
I've also started on making the encapsulation better. A lot of the scripts have public variables, meaning anything can access them. We got points docked off for this, so I'm making them private and adding getters + setters in their place. It clogs up the bottom of the scripts, but I suppose that's the price to pay for better encapsulation. (There's probably a better way of achieving what I'm trying to do, but I can't think of it.)