rearranged player shooting scripts
This update is a precursor to adding limited ammo to the player. I wanted to move the shooting scripts into a new child, that way everything would be around the same place.
(Also, I did export and test it before submitting. We're not going to have the shooting glitch like last time.)
(I'm also submitting this precursor in case there are weird glitches in the ammo update.)