Added Player AI (Incomplete)
Added some scripts, I'm pretty sure we can refactor it a lot, but I just added them for reference.
The enemy detection works well, and the targeting is working fine too.
There are two problems for now, one is that the enemy is not getting the player's coordinates to follow, the other one is the AI bullets don't damage the enemies.