Added enemy navmesh and player AI spawning location
I implemented the last agent behaviors for the enemies. Also, the regular enemy will flee the scene hen reaching low health and stop shooting.
The player AI will spawn in the 2nd location and follow the player around the map, but will attack enemies that enter its view.
I'm not sure why is not damaging enemies