added damaged "animation"
Per the request of one of the playtesters, I've added a damaging animation, where the entity damaged (player, enemy, shield, etc.) flashes red for a brief moment.
I've also sorted out the explosive enemy's attack script and connected it to the base shooting script, so that the detector will work properly.
Currently, the console is outputting an error involving the player AI. The UI manager is trying to display the status of the AI's shield child - but it doesn't have one. Should the AI have access to powerups? Also, should the AI have unlimited ammo? (The ammo display doesn't update when the AI shoots.)