added text outline to dialog scene; added fading animation; fixed canvases
I added an outline to the text in the dialog scene (both to the text and the nametags).
I also added a fading animation, where, when the active speaker switches, the active speaker "fades" brighter, while the inactive speaker "fades" darker.
Also, apparently, to make resolutions in the editor appear as they should in other resolutions, you need to change the UI Scale Mode in the Canvas Scalar to "Scale with Screen Size".