refactoured boss scripts; added win scene; added file reading for dialogue
I refactoured the boss scripts to have better encapsulation, and added a win scene that appears after the boss is defeated.
I also implemented dialogue reading; however, it reads TXT files, not JSON. All text files must be placed under Assets/Dialogue in order to be usable. There are some instructions on how to format dialogue scripts in the dialog_test file. You also have to manually attach the file to the dialogue manager via the inspector in the editor.