changed title art pan; fixed dialogue screen loading
I changed the title art's pan from logarithmic to linear; it was causing problems in the build.
I also fixed the dialogue screen loading, but there's a problem when building it. All dialogue has to be copied and pasted manually into the "Assets/Dialogue" folder. This is important for the in-game cutscene to play.