added testing shooting-teleporter room
I've added a teleport flag system, where certain conditions can be added before a teleporter is usable. I've also added a scene experimenting with this: the player must shoot all the targets before the teleporter can activate.
However, there are two issues. First, if you shoot on the teleporter when you first load in (without moving), the bullets don't collide with the targets, but do collide with the walls. Second, if the player completes their task while standing on top of the teleporter, they won't be able to use it until they step off first.
I've also created a simple camera; this should preferably only be used in singleplayer.